Crate openrr_config

OpenRR Config

Utilities for modifying configuration files.

License

Licensed under the Apache License, Version 2.0.

Functions

evaluate

Evaluates the given string and returns a concatenated string of the results.

overwrite

Replaces the contents of the specified TOML document based on the specified scripts.

overwrite_str

Replaces the contents of the specified TOML document based on the specified scripts, returning edited document as string.